+++ hackage-security/ChangeLog.md
@@ -1,7 +1,7 @@
 unreleased
 ----------
 
-* Allow `network-2.7.0.0`
+* Allow `network-3.0.0.0`
 * Allow `aeson-1.4.0.0`
 
 0.5.3.0